switch to monthly releases
Previously, it was set to a daily release (see the cron expresseion)
for testing. That's been running fine for a while now, so we can
switch to only monthly releases. Furthermore, the seconds portion
of the tag can be removed as it was only there to prevent tag
conflicts from the daily releases.